requirements.txt
bleak
fastapi
uvicorn
websockets
pyyaml
influxdb
paho-mqtt
python-dotenv
async-timeout